+++ packages/web/dates/src/models/dateFormat.model.ts
@@ -1,0 +1,8 @@
+export type DateFormat
+  = | 'de-DE'
+    | 'en-GB'
+    | 'en-US'
+    | 'ja-JP'
+    | 'locale-default'
+    | 'nl-NL'
+    | 'sv-SE'